Fix potential overflow in combined ESC sensor data

+++ src/main/sensors/esc_sensor.h
@@ -27,11 +27,11 @@ PG_DECLARE(escSensorConfig_t, escSensorConfig);
 
 typedef struct {
     uint8_t dataAge;
-    int8_t temperature;
-    int16_t voltage;
-    int16_t current;
-    int16_t consumption;
-    int16_t rpm;
+    int8_t temperature;  // C degrees
+    int16_t voltage;     // 0.01V
+    int32_t current;     // 0.01A
+    int32_t consumption; // mAh
+    int16_t rpm;         // 0.01erpm
 } escSensorData_t;
 
 #define ESC_DATA_INVALID 255
